requirements.txt
streamlit
python-dotenv
google-generativeai
gTTS
SpeechRecognition
pydub